+++ lib/blocs/attendance/attendance_state.dart
@@ -24,7 +24,7 @@ class AttendanceState extends Equatable {
   final DateTime clockout;
   final DateTime workDate;
   final int status;
-  final int hourlyRate;
+  final double hourlyRate;
   final AttendanceStatus attStatus;
   final List<Attendance> attendanceList;
   final CustomError customError;
@@ -55,7 +55,7 @@ class AttendanceState extends Equatable {
       clockout: DateTime(1970, 1, 1),
       workDate: DateTime(1970, 1, 1),
       status: 0,
-      hourlyRate: 0,
+      hourlyRate: 0.0,
       attStatus: AttendanceStatus.initial,
       attendanceList: const [],
       customError: const CustomError(),
@@ -89,7 +89,7 @@ class AttendanceState extends Equatable {
     DateTime? clockout,
     DateTime? workDate,
     int? status,
-    int? hourlyRate,
+    double? hourlyRate,
     AttendanceStatus? attStatus,
     List<Attendance>? attendanceList,
     CustomError? customError,
